Implementing room database

+++ app/src/main/java/com/example/android_mvvm_best_pratices/ui/component/base/BaseAdapter.kt
@@ -53,5 +53,5 @@ class BaseViewHolder<BINDING : ViewDataBinding>(val binder: BINDING) :
     RecyclerView.ViewHolder(binder.root)
 
 interface AdapterListItem {
-    val id: Long?
+    val id: Int?
 }
